+++ packages/orq-rc/RELEASES.md
@@ -11259,3 +11259,13 @@ Based on:
 - [typescript v4.13.0-rc.17] packages/orq-rc
 ### Releases
 - [NPM v4.13.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.17 - packages/orq-rc
+
+## 2026-07-30 05:48:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.791.0 (2.924.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.13.0-rc.18] packages/orq-rc
+### Releases
+- [NPM v4.13.0-rc.18] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.18 - packages/orq-rc